+++ src/wallpaper_generators/solid_color.rs
@@ -1,10 +1,92 @@
-use super::super::cli::{Config, Mode, SolidMode};
-use super::super::os_implementations::get_screen_resolution;
+use super::super::{cli::SolidMode, config::Config, os_implementations::get_screen_resolution};
 use super::utils::{AstraImage, WallpaperGeneratorError};
 use clap::ValueEnum;
 use image::{ImageBuffer, Rgb};
+use rand::{Rng, rng};
+use serde::Deserialize;
 
-#[derive(Copy, Clone, Debug, ValueEnum)]
+pub fn generate_solid_color(
+    config: &Config,
+    mode: &SolidMode,
+) -> Result<AstraImage, WallpaperGeneratorError> {
+    config.print_if_verbose("Generating solid color image...");
+
+    let (width, height) =
+        get_screen_resolution().map_err(|e| WallpaperGeneratorError::OSError(e.to_string()))?;
+    config.print_if_verbose(format!("Detected screen resolution: {}x{}", width, height).as_str());
+
+    if config.respect_user_config {
+        config.print_if_verbose("User config detected with solid_gen options...");
+    }
+    // TODO: v1.1.0 - implement color theme logic will need to make ThemeSelector from config
+    // let theme = crate::respect_user_config_or_default!(config, solid_gen, respect_color_themes, { ThemeSelector::random() })?;
+
+    let mut mode_options: Vec<SolidMode> = vec![];
+    let preferred_default_colors: Vec<Color> =
+        crate::respect_user_config_or_default!(config, solid_gen, preferred_default_colors, {
+            Ok(vec![])
+        })?;
+    let preferred_rgb_colors: Vec<(u8, u8, u8)> =
+        crate::respect_user_config_or_default!(config, solid_gen, preferred_rgb_colors, {
+            Ok(vec![])
+        })?;
+
+    preferred_default_colors
+        .iter()
+        .for_each(|color| mode_options.push(SolidMode::Color { name: *color }));
+    preferred_rgb_colors.iter().for_each(|(r, g, b)| {
+        mode_options.push(SolidMode::Rgb {
+            r: *r,
+            g: *g,
+            b: *b,
+        })
+    });
+
+    let imgbuf: AstraImage;
+    if mode_options.is_empty() {
+        config.print_if_verbose(
+            "read preferred_default_colors & preferred_rgb_colors config, but none were found",
+        );
+        // Use mode passed in instead since no config setup
+        imgbuf = generate_image(&mode, width, height);
+    } else {
+        config.print_if_verbose(
+            "selecting random mode based on preferred_default_colors & preferred_rgb_colors config",
+        );
+        let mut rng = rng();
+        let n = rng.random_range(..mode_options.len());
+        let rand_mode = mode_options
+            .get(n)
+            .expect("random selected solid mode from user config should be defined");
+        imgbuf = generate_image(rand_mode, width, height);
+    }
+
+    config.print_if_verbose("Image generated!");
+    Ok(imgbuf)
+}
+
+fn generate_image(mode: &SolidMode, width: u32, height: u32) -> AstraImage {
+    match mode {
+        SolidMode::Random => ImageBuffer::from_pixel(
+            width,
+            height,
+            Rgb([
+                rand::random::<u8>(),
+                rand::random::<u8>(),
+                rand::random::<u8>(),
+            ]),
+        ),
+        SolidMode::Rgb { r, g, b } => {
+            ImageBuffer::from_pixel(width, height, Rgb([r.clone(), g.clone(), b.clone()]))
+        }
+        SolidMode::Color { name } => {
+            let (r, g, b) = name.rgb();
+            ImageBuffer::from_pixel(width, height, Rgb([r, g, b]))
+        }
+    }
+}
+
+#[derive(Copy, Clone, Debug, Deserialize, PartialEq, ValueEnum)]
 pub enum Color {
     White,
     Black,
@@ -53,51 +135,4 @@ impl Color {
             Color::Beige => (245, 245, 220),
         }
     }
-}
-
-pub fn generate_solid_color(
-    config: &Config,
-    mode: Option<Mode>,
-) -> Result<AstraImage, WallpaperGeneratorError> {
-    config.print_if_verbose("Generating solid color...");
-
-    let (width, height) =
-        get_screen_resolution().map_err(|e| WallpaperGeneratorError::OSError(e.to_string()))?;
-    config.print_if_verbose(format!("Detected screen resolution: {}x{}", width, height).as_str());
-
-    config.print_if_verbose("Generating image...");
-
-    let mode = mode.ok_or(WallpaperGeneratorError::NoModeProvided(
-        "Solid color generator requires a SolidMode".to_string(),
-    ))?;
-    // Expect a SolidMode, error is other mode provided
-    let solid_mode: SolidMode = match mode {
-        Mode::Solid(mode) => mode,
-        // Leave this in as in the future more modes can be added
-        _ => {
-            return Err(WallpaperGeneratorError::InvalidMode(
-                "SolidMode".to_string(),
-            ));
-        }
-    };
-    let imgbuf = match solid_mode {
-        SolidMode::Random => ImageBuffer::from_pixel(
-            width,
-            height,
-            Rgb([
-                rand::random::<u8>(),
-                rand::random::<u8>(),
-                rand::random::<u8>(),
-            ]),
-        ),
-        SolidMode::Rgb { r, g, b } => ImageBuffer::from_pixel(width, height, Rgb([r, g, b])),
-        SolidMode::Color { name } => {
-            let (r, g, b) = name.rgb();
-            ImageBuffer::from_pixel(width, height, Rgb([r, g, b]))
-        }
-    };
-
-    config.print_if_verbose("Image generated!");
-
-    Ok(imgbuf)
 }